fix(ios): Crash "Unexpectedly found nil while implicitly unwrapping an Optional value"
The crash was occuring inside `render` method in `Map.swift`. The creation of `GMapView` sometimes was occuring after trying to acces it as `self.mapViewController.GMapView`. In such case the value was still nil, resulting in the crash when trying to unwrap the value.

The fix involved only accessing `self.mapViewController.GMapView` after `viewDidLoad` finishes and `GMapView` is initialized.